♻️ refactor(monitoring): extract codex execution workflow
Move Codex inspection delete, disable, enable, and refresh execution into a dedicated model module.

Keep the public codexInspection export stable so page and tests continue using the same API.

Validated with type-check, focused monitoring tests, lint, and Vite production build.
